moved resolve_locale into the new locale module
