+++ leptos_i18n/src/context.rs
@@ -121,7 +121,7 @@ pub type CookieOptions<L> = UseCookieOptions<
     <FromToStringCodec as codee::Decoder<L>>::Error,
 >;
 
-const ENABLE_COOKIE: bool = cfg!(feature = "cookie");
+pub(crate) const ENABLE_COOKIE: bool = cfg!(feature = "cookie");
 
 const COOKIE_PREFERED_LANG: &str = "i18n_pref_locale";
 
@@ -167,14 +167,14 @@ where
     L: Locale,
 {
     /// Should set a cookie to keep track of the locale when page reload (default to true) (do nothing without the "cookie" feature)
-    enable_cookie: bool,
+    pub enable_cookie: bool,
     /// Give a custom name to the cookie (default to the crate default value) (do nothing without the "cookie" feature or if `enable_cookie` is false)
     #[builder(into)]
-    cookie_name: Cow<'a, str>,
+    pub cookie_name: Cow<'a, str>,
     /// Options for the cookie, the value is of type `leptos_use::UseCookieOptions<Locale>` (default to `Default::default`)
-    cookie_options: CookieOptions<L>,
+    pub cookie_options: CookieOptions<L>,
     /// Options to pass to `leptos_use::use_locales`.
-    ssr_lang_header_getter: UseLocalesOptions,
+    pub ssr_lang_header_getter: UseLocalesOptions,
 }
 
 impl<L: Locale> Default for I18nContextOptions<'_, L> {
@@ -186,41 +186,6 @@ impl<L: Locale> Default for I18nContextOptions<'_, L> {
             ssr_lang_header_getter: Default::default(),
         }
     }
-}
-
-/// Same as `resolve_locale` but with some cookies options.
-pub fn resolve_locale_with_options<L: Locale>(options: I18nContextOptions<L>) -> L {
-    let I18nContextOptions {
-        enable_cookie,
-        cookie_name,
-        cookie_options,
-        ssr_lang_header_getter,
-    } = options;
-    let (lang_cookie, _) = if ENABLE_COOKIE && enable_cookie {
-        leptos_use::use_cookie_with_options::<L, FromToStringCodec>(&cookie_name, cookie_options)
-    } else {
-        let (lang_cookie, set_lang_cookie) = signal(None);
-        (lang_cookie.into(), set_lang_cookie)
-    };
-    fetch_locale::resolve_locale(lang_cookie.get_untracked(), ssr_lang_header_getter)
-}
-
-/// Resolve the locale.
-///
-/// This as the same behavior as calling `init_i18n_context().get_locale_untracked()`.
-///
-/// This function primary usage is to access a user locale in a server function, but is not constrained to it.
-///
-/// Here is the list of detection methods, sorted in priorities:
-/// 1. The "lang" attribute is set on the `<html>` element in hydrate
-/// 1. A cookie is present that contains a previously detected locale
-/// 1. A locale can be matched based on the [`Accept-Language` header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Language) in SSR
-/// 1. A locale can be matched based on the [`navigator.languages` API](https://developer.mozilla.org/en-US/docs/Web/API/Navigator/languages) in CSR
-/// 1. As a last resort, the default locale is used.
-///
-/// *note*: this function does not take into account URL locale prefix when using `I18nRoute` (e.g. `/en/about`)
-pub fn resolve_locale<L: Locale>() -> L {
-    resolve_locale_with_options(Default::default())
 }
 
 /// Same as `init_i18n_context` but with some cookies options.
